+++ eventsapp/uix/cards/__init__.py
@@ -4,3 +4,5 @@ from .talkdetailcard import TalkDetailCard
 from .sponsordetailcard import SponsorDetailCard
 
 from .cardscontainer import CardsContainer
+from .cardscontainer import CardBoxLayout
+from .cardscontainer import CardStackLayout
