+++ eventsapp/uix/screens/openspacescreen.py
@@ -6,6 +6,7 @@ Open Space Screen:
 from kivy.uix.screenmanager import Screen
 from kivy.lang import Builder
 from uix.buttons import SocialButton
+from uix.cards import CardsContainer, CardBoxLayout
 import json
 
 
@@ -17,56 +18,49 @@ class OpenSpaceScreen(Screen):
     BoxLayout
         orientation: 'vertical'
         TopBar
-        BoxLayout:
-            orientation: 'vertical'
-            padding: dp(20), dp(10)
-            spacing: dp(40)
-            canvas:
-                Color:
-                    rgba: (.91, .91, .91, 1)
-                Rectangle:
-                    pos: self.pos
-                    size: self.size
-            AsyncImage
-                id: logo
-                size_hint: 1, .7
-                source: 'data/images/logo.png'
-            BoxLayout:
-                canvas.before:
-                    Color:
-                        rgba: (.81, .81, .81, 1)
-                    Rectangle:
-                        pos: self.pos
-                        size: self.size
-                orientation: 'vertical'
-                padding: dp(5), dp(20)
+        CardsContainer:
+            size_hint_y: 1
+            CardBoxLayout:
+                AsyncImage
+                    id: logo
+                    size_hint: 1, .7
+                    source: 'data/images/logo.png'
+                BoxLayout:
+                    canvas.before:
+                        Color:
+                            rgba: (.81, .81, .81, 1)
+                        Rectangle:
+                            pos: self.pos
+                            size: self.size
+                    orientation: 'vertical'
+                    padding: dp(5), dp(20)
+                    Label:
+                        id: quote
+                        text: ''
+                        text_size: self.size
+                        halign: 'center'
+                        valign: 'top'
+                        font_size: dp(20)
+                        bold: True
+                        color: 0, 0, 0, 1
+                    Label:
+                        size_hint: 1, .5
+                        id: author
+                        text: 'Michael M Pannwitz, Open Space practitioner'
+                        text_size: self.size
+                        halign: 'center'
+                        valign: 'center'
+                        font_size: dp(20)
+                        bold: True
+                        color: 0, 0, 0, 1
                 Label:
-                    id: quote
+                    id: about
                     text: ''
                     text_size: self.size
-                    halign: 'center'
+                    halign: 'left'
                     valign: 'top'
                     font_size: dp(20)
-                    bold: True
                     color: 0, 0, 0, 1
-                Label:
-                    size_hint: 1, .5
-                    id: author
-                    text: 'Michael M Pannwitz, Open Space practitioner'
-                    text_size: self.size
-                    halign: 'center'
-                    valign: 'center'
-                    font_size: dp(20)
-                    bold: True
-                    color: 0, 0, 0, 1
-            Label:
-                id: about
-                text: ''
-                text_size: self.size
-                halign: 'left'
-                valign: 'top'
-                font_size: dp(20)
-                color: 0, 0, 0, 1
     ''')
 
     def on_pre_enter(self):
